Move spriteParser to graphics
