+++ src/ecs/entities/dandelionFactory.ts
@@ -3,7 +3,7 @@ import { Entity } from '../entity'
 import { PositionComponent } from '../components/positionComponent'
 import { DrawComponent } from '../components/drawComponent'
 import { AIComponent } from '../components/aiComponent'
-import { parseSprite } from '../../parser/spriteParser'
+import { parseSprite } from '../../graphics/spriteParser'
 import { AnimationStateComponent } from '../components/animationStateComponent'
 import dandelionDefinition from '../../../res/entities/dandelion.json'
 import { World } from '../world'
