+++ src/ecs/entities/dandelionFluffFactory.ts
@@ -6,7 +6,7 @@ import { DrawComponent } from '../components/drawComponent'
 import { ColliderComponent, AABBDef } from '../components/colliderComponent'
 import { CategoryList } from './category'
 import { AIComponent } from '../components/aiComponent'
-import { parseSprite } from '../../parser/spriteParser'
+import { parseSprite } from '../../graphics/spriteParser'
 import { AnimationStateComponent } from '../components/animationStateComponent'
 import { PickupTargetComponent } from '../components/pickupTargetComponent'
 import dandelionFluffDefinition from '../../../res/entities/dandelion_fluff.json'
