+++ src/ecs/entities/enemy1Factory.ts
@@ -11,7 +11,7 @@ import { AttackComponent } from '../components/attackComponent'
 import { HPComponent } from '../components/hpComponent'
 import { InvincibleComponent } from '../components/invincibleComponent'
 import { AIComponent } from '../components/aiComponent'
-import { parseSprite } from '../../parser/spriteParser'
+import { parseSprite } from '../../graphics/spriteParser'
 import { AnimationStateComponent } from '../components/animationStateComponent'
 import enemy1Definition from '../../../res/entities/enemy1.json'
 import { World } from '../world'
