+++ src/ecs/entities/mossFactory.ts
@@ -2,7 +2,7 @@ import { Entity } from '../entity'
 import { EntityFactory } from './entityFactory'
 import { PositionComponent } from '../components/positionComponent'
 import { DrawComponent } from '../components/drawComponent'
-import { parseSprite } from '../../parser/spriteParser'
+import { parseSprite } from '../../graphics/spriteParser'
 import mossDefinition from '../../../res/entities/moss.json'
 import { World } from '../world'
 import { AABBDef, ColliderComponent } from '../components/colliderComponent'
